+++ lib/src/controls/utils/scrollbar.dart
@@ -63,13 +63,10 @@ class _ScrollbarState extends RawScrollbarState<Scrollbar> {
   }
 
   Set<WidgetState> get _currentState {
-    if (_dragIsActive) {
-      return {WidgetState.pressed};
-    } else if (_hoverIsActive) {
-      return {WidgetState.hovered};
-    } else {
-      return {};
-    }
+    return {
+      if (_dragIsActive) WidgetState.pressed,
+      if (_hoverIsActive) WidgetState.hovered,
+    };
   }
 
   Color _trackColor(Set<WidgetState> state) {
@@ -142,6 +139,7 @@ class _ScrollbarState extends RawScrollbarState<Scrollbar> {
     super.handleThumbPressStart(localPosition);
     if (mounted) {
       setState(() => _dragIsActive = true);
+      _updateAnimation();
     }
   }
 
@@ -150,6 +148,7 @@ class _ScrollbarState extends RawScrollbarState<Scrollbar> {
     super.handleThumbPressEnd(localPosition, velocity);
     if (mounted) {
       setState(() => _dragIsActive = false);
+      _updateAnimation();
     }
   }
 
@@ -162,14 +161,14 @@ class _ScrollbarState extends RawScrollbarState<Scrollbar> {
       await contractDelay;
       if (mounted) {
         setState(() => _hoverIsActive = true);
-        _hoverAnimationController.forward();
+        _updateAnimation();
       }
     } else if (_hoverIsActive) {
       await contractDelay;
       if (mounted) {
         // Pointer was, but is no longer over painted scrollbar.
-        await _hoverAnimationController.reverse();
         setState(() => _hoverIsActive = false);
+        _updateAnimation();
       }
     }
   }
@@ -179,7 +178,7 @@ class _ScrollbarState extends RawScrollbarState<Scrollbar> {
     super.handleHoverExit(event);
     if (mounted) {
       setState(() => _hoverIsActive = false);
-      _hoverAnimationController.reverse();
+      _updateAnimation();
     }
   }
 
@@ -187,6 +186,14 @@ class _ScrollbarState extends RawScrollbarState<Scrollbar> {
   void dispose() {
     _hoverAnimationController.dispose();
     super.dispose();
+  }
+
+  void _updateAnimation() {
+    if (_currentState.isEmpty) {
+      _hoverAnimationController.reverse();
+    } else {
+      _hoverAnimationController.forward();
+    }
   }
 }